pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cam"
version = "0.2.0"
description = "AI's camera - Multi-desktop screenshot capture with MCP server for AI agents"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "Yusuke Watanabe", email = "Yusuke.Watanabe@scitex.ai"}
]
keywords = ["screenshot", "screen-capture", "wsl", "windows", "monitoring", "automation", "ai-vision"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Multimedia :: Graphics :: Capture :: Screen Capture",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
]

# Core dependencies
dependencies = [
    "fastmcp>=2.0.0",
    "mss>=6.0.0",
    "Pillow>=9.0.0",
]

[project.optional-dependencies]
browser = [
    "playwright>=1.40.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
cam = "cam.cli:main"

[project.urls]
Homepage = "https://github.com/ywatanabe1989/cam"
Repository = "https://github.com/ywatanabe1989/cam"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py37", "py38", "py39", "py310", "py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
